+++ ConstraidTests/CenterWithinEdgesTests.swift
@@ -2,13 +2,13 @@ import XCTest
 import Constraid
 
 class CenterWithinEdgesTests: XCTestCase {
-    func testCenterVerticallyWithin() {
+    func testCenterVerticallyWithinOffsetDown() {
         let viewOne = UIView()
         let viewTwo = UIView()
 
         viewOne.addSubview(viewTwo)
 
-        let constraints = viewOne.center(verticallyWithin: viewTwo, constant: 10.0, multiplier: 2.0, priority: 500)
+        let constraints = viewOne.center(verticallyWithin: viewTwo, offsetBy: 10.0, offsetDirection: .down, multiplier: 2.0, priority: 500)
         let constraintOne = viewOne.constraints.first! as NSLayoutConstraint
 
         XCTAssertEqual(constraints, viewOne.constraints)
@@ -23,13 +23,34 @@ class CenterWithinEdgesTests: XCTestCase {
         XCTAssertEqual(constraintOne.priority, 500)
     }
 
-    func testCenterHorizontallyWithin() {
+    func testCenterVerticallyWithinOffsetUp() {
         let viewOne = UIView()
         let viewTwo = UIView()
 
         viewOne.addSubview(viewTwo)
 
-        let constraints = viewOne.center(horizontallyWithin: viewTwo, constant: 10.0, multiplier: 2.0, priority: 500)
+        let constraints = viewOne.center(verticallyWithin: viewTwo, offsetBy: 10.0, offsetDirection: .up, multiplier: 2.0, priority: 500)
+        let constraintOne = viewOne.constraints.first! as NSLayoutConstraint
+
+        XCTAssertEqual(constraints, viewOne.constraints)
+        XCTAssertEqual(constraintOne.isActive, true)
+        XCTAssertEqual(constraintOne.firstItem as! UIView, viewOne)
+        XCTAssertEqual(constraintOne.firstAttribute, NSLayoutAttribute.centerY)
+        XCTAssertEqual(constraintOne.relation, NSLayoutRelation.equal)
+        XCTAssertEqual(constraintOne.secondItem as! UIView, viewTwo)
+        XCTAssertEqual(constraintOne.secondAttribute, NSLayoutAttribute.centerY)
+        XCTAssertEqual(constraintOne.constant, -10.0)
+        XCTAssertEqual(constraintOne.multiplier, 2.0)
+        XCTAssertEqual(constraintOne.priority, 500)
+    }
+
+    func testCenterHorizontallyWithinOffsetRight() {
+        let viewOne = UIView()
+        let viewTwo = UIView()
+
+        viewOne.addSubview(viewTwo)
+
+        let constraints = viewOne.center(horizontallyWithin: viewTwo, offsetBy: 10.0, offsetDirection: .right, multiplier: 2.0, priority: 500)
         let constraintOne = viewOne.constraints.first! as NSLayoutConstraint
 
         XCTAssertEqual(constraints, viewOne.constraints)
@@ -44,13 +65,34 @@ class CenterWithinEdgesTests: XCTestCase {
         XCTAssertEqual(constraintOne.priority, 500)
     }
 
-    func testCenterWithin() {
+    func testCenterHorizontallyWithinOffsetLeft() {
         let viewOne = UIView()
         let viewTwo = UIView()
 
         viewOne.addSubview(viewTwo)
 
-        let constraints = viewOne.center(within: viewTwo, constant: 10.0, multiplier: 2.0, priority: 500)
+        let constraints = viewOne.center(horizontallyWithin: viewTwo, offsetBy: 10.0, offsetDirection: .left, multiplier: 2.0, priority: 500)
+        let constraintOne = viewOne.constraints.first! as NSLayoutConstraint
+
+        XCTAssertEqual(constraints, viewOne.constraints)
+        XCTAssertEqual(constraintOne.isActive, true)
+        XCTAssertEqual(constraintOne.firstItem as! UIView, viewOne)
+        XCTAssertEqual(constraintOne.firstAttribute, NSLayoutAttribute.centerX)
+        XCTAssertEqual(constraintOne.relation, NSLayoutRelation.equal)
+        XCTAssertEqual(constraintOne.secondItem as! UIView, viewTwo)
+        XCTAssertEqual(constraintOne.secondAttribute, NSLayoutAttribute.centerX)
+        XCTAssertEqual(constraintOne.constant, -10.0)
+        XCTAssertEqual(constraintOne.multiplier, 2.0)
+        XCTAssertEqual(constraintOne.priority, 500)
+    }
+
+    func testCenterWithinOffsetDownAndRight() {
+        let viewOne = UIView()
+        let viewTwo = UIView()
+
+        viewOne.addSubview(viewTwo)
+
+        let constraints = viewOne.center(within: viewTwo, offsetBy: 10.0, offsetDirection: .downAndRight, multiplier: 2.0, priority: 500)
 
         let constraintOne = viewOne.constraints.first! as NSLayoutConstraint
         let constraintTwo = viewOne.constraints.last! as NSLayoutConstraint
@@ -74,6 +116,40 @@ class CenterWithinEdgesTests: XCTestCase {
         XCTAssertEqual(constraintTwo.secondItem as! UIView, viewTwo)
         XCTAssertEqual(constraintTwo.secondAttribute, NSLayoutAttribute.centerY)
         XCTAssertEqual(constraintTwo.constant, 10.0)
+        XCTAssertEqual(constraintTwo.multiplier, 2.0)
+        XCTAssertEqual(constraintTwo.priority, 500)
+    }
+
+    func testCenterWithinOffsetUpAndLeft() {
+        let viewOne = UIView()
+        let viewTwo = UIView()
+
+        viewOne.addSubview(viewTwo)
+
+        let constraints = viewOne.center(within: viewTwo, offsetBy: 10.0, offsetDirection: .upAndLeft, multiplier: 2.0, priority: 500)
+
+        let constraintOne = viewOne.constraints.first! as NSLayoutConstraint
+        let constraintTwo = viewOne.constraints.last! as NSLayoutConstraint
+
+        XCTAssertEqual(constraints, viewOne.constraints)
+
+        XCTAssertEqual(constraintOne.isActive, true)
+        XCTAssertEqual(constraintOne.firstItem as! UIView, viewOne)
+        XCTAssertEqual(constraintOne.firstAttribute, NSLayoutAttribute.centerX)
+        XCTAssertEqual(constraintOne.relation, NSLayoutRelation.equal)
+        XCTAssertEqual(constraintOne.secondItem as! UIView, viewTwo)
+        XCTAssertEqual(constraintOne.secondAttribute, NSLayoutAttribute.centerX)
+        XCTAssertEqual(constraintOne.constant, -10.0)
+        XCTAssertEqual(constraintOne.multiplier, 2.0)
+        XCTAssertEqual(constraintOne.priority, 500)
+
+        XCTAssertEqual(constraintTwo.isActive, true)
+        XCTAssertEqual(constraintTwo.firstItem as! UIView, viewOne)
+        XCTAssertEqual(constraintTwo.firstAttribute, NSLayoutAttribute.centerY)
+        XCTAssertEqual(constraintTwo.relation, NSLayoutRelation.equal)
+        XCTAssertEqual(constraintTwo.secondItem as! UIView, viewTwo)
+        XCTAssertEqual(constraintTwo.secondAttribute, NSLayoutAttribute.centerY)
+        XCTAssertEqual(constraintTwo.constant, -10.0)
         XCTAssertEqual(constraintTwo.multiplier, 2.0)
         XCTAssertEqual(constraintTwo.priority, 500)
     }