+++ ConstraidTests/CenterWithinMarginsTests.swift
@@ -2,13 +2,13 @@ import XCTest
 import Constraid
 
 class CenterWithinMarginsTests: XCTestCase {
-    func testCenterVerticallyWithinMargins() {
+    func testCenterVerticallyWithinMarginsOffsetDown() {
         let viewOne = UIView()
         let viewTwo = UIView()
 
         viewOne.addSubview(viewTwo)
 
-        let constraints = viewOne.center(verticallyWithinMarginsOf: viewTwo, constant: 10.0, multiplier: 2.0, priority: 500)
+        let constraints = viewOne.center(verticallyWithinMarginsOf: viewTwo, offsetBy: 10.0, offsetDirection: .down, multiplier: 2.0, priority: 500)
         let constraintOne = viewOne.constraints.first! as NSLayoutConstraint
 
         XCTAssertEqual(constraints, viewOne.constraints)
@@ -23,13 +23,34 @@ class CenterWithinMarginsTests: XCTestCase {
         XCTAssertEqual(constraintOne.priority, 500)
     }
 
-    func testCenterHorizontallyWithinMargins() {
+    func testCenterVerticallyWithinMarginsOffsetUp() {
         let viewOne = UIView()
         let viewTwo = UIView()
 
         viewOne.addSubview(viewTwo)
 
-        let constraints = viewOne.center(horizontallyWithinMarginsOf: viewTwo, constant: 10.0, multiplier: 2.0, priority: 500)
+        let constraints = viewOne.center(verticallyWithinMarginsOf: viewTwo, offsetBy: 10.0, offsetDirection: .up, multiplier: 2.0, priority: 500)
+        let constraintOne = viewOne.constraints.first! as NSLayoutConstraint
+
+        XCTAssertEqual(constraints, viewOne.constraints)
+        XCTAssertEqual(constraintOne.isActive, true)
+        XCTAssertEqual(constraintOne.firstItem as! UIView, viewOne)
+        XCTAssertEqual(constraintOne.firstAttribute, NSLayoutAttribute.centerY)
+        XCTAssertEqual(constraintOne.relation, NSLayoutRelation.equal)
+        XCTAssertEqual(constraintOne.secondItem as! UIView, viewTwo)
+        XCTAssertEqual(constraintOne.secondAttribute, NSLayoutAttribute.centerYWithinMargins)
+        XCTAssertEqual(constraintOne.constant, -10.0)
+        XCTAssertEqual(constraintOne.multiplier, 2.0)
+        XCTAssertEqual(constraintOne.priority, 500)
+    }
+
+    func testCenterHorizontallyWithinMarginsOffsetRight() {
+        let viewOne = UIView()
+        let viewTwo = UIView()
+
+        viewOne.addSubview(viewTwo)
+
+        let constraints = viewOne.center(horizontallyWithinMarginsOf: viewTwo, offsetBy: 10.0, offsetDirection: .right, multiplier: 2.0, priority: 500)
         let constraintOne = viewOne.constraints.first! as NSLayoutConstraint
 
         XCTAssertEqual(constraints, viewOne.constraints)
@@ -44,13 +65,34 @@ class CenterWithinMarginsTests: XCTestCase {
         XCTAssertEqual(constraintOne.priority, 500)
     }
 
-    func testCenterWithinMargins() {
+    func testCenterHorizontallyWithinMarginsOffsetLeft() {
         let viewOne = UIView()
         let viewTwo = UIView()
 
         viewOne.addSubview(viewTwo)
 
-        let constraints = viewOne.center(withinMarginsOf: viewTwo, constant: 10.0, multiplier: 2.0, priority: 500)
+        let constraints = viewOne.center(horizontallyWithinMarginsOf: viewTwo, offsetBy: 10.0, offsetDirection: .left, multiplier: 2.0, priority: 500)
+        let constraintOne = viewOne.constraints.first! as NSLayoutConstraint
+
+        XCTAssertEqual(constraints, viewOne.constraints)
+        XCTAssertEqual(constraintOne.isActive, true)
+        XCTAssertEqual(constraintOne.firstItem as! UIView, viewOne)
+        XCTAssertEqual(constraintOne.firstAttribute, NSLayoutAttribute.centerX)
+        XCTAssertEqual(constraintOne.relation, NSLayoutRelation.equal)
+        XCTAssertEqual(constraintOne.secondItem as! UIView, viewTwo)
+        XCTAssertEqual(constraintOne.secondAttribute, NSLayoutAttribute.centerXWithinMargins)
+        XCTAssertEqual(constraintOne.constant, -10.0)
+        XCTAssertEqual(constraintOne.multiplier, 2.0)
+        XCTAssertEqual(constraintOne.priority, 500)
+    }
+
+    func testCenterWithinMarginsOffsetDownAndRight() {
+        let viewOne = UIView()
+        let viewTwo = UIView()
+
+        viewOne.addSubview(viewTwo)
+
+        let constraints = viewOne.center(withinMarginsOf: viewTwo, offsetBy: 10.0, offsetDirection: .downAndRight, multiplier: 2.0, priority: 500)
 
         let constraintOne = viewOne.constraints.first! as NSLayoutConstraint
         let constraintTwo = viewOne.constraints.last! as NSLayoutConstraint
@@ -74,6 +116,40 @@ class CenterWithinMarginsTests: XCTestCase {
         XCTAssertEqual(constraintTwo.secondItem as! UIView, viewTwo)
         XCTAssertEqual(constraintTwo.secondAttribute, NSLayoutAttribute.centerYWithinMargins)
         XCTAssertEqual(constraintTwo.constant, 10.0)
+        XCTAssertEqual(constraintTwo.multiplier, 2.0)
+        XCTAssertEqual(constraintTwo.priority, 500)
+    }
+
+    func testCenterWithinMarginsOffsetUpAndRight() {
+        let viewOne = UIView()
+        let viewTwo = UIView()
+
+        viewOne.addSubview(viewTwo)
+
+        let constraints = viewOne.center(withinMarginsOf: viewTwo, offsetBy: 10.0, offsetDirection: .upAndLeft, multiplier: 2.0, priority: 500)
+
+        let constraintOne = viewOne.constraints.first! as NSLayoutConstraint
+        let constraintTwo = viewOne.constraints.last! as NSLayoutConstraint
+
+        XCTAssertEqual(constraints, viewOne.constraints)
+
+        XCTAssertEqual(constraintOne.isActive, true)
+        XCTAssertEqual(constraintOne.firstItem as! UIView, viewOne)
+        XCTAssertEqual(constraintOne.firstAttribute, NSLayoutAttribute.centerX)
+        XCTAssertEqual(constraintOne.relation, NSLayoutRelation.equal)
+        XCTAssertEqual(constraintOne.secondItem as! UIView, viewTwo)
+        XCTAssertEqual(constraintOne.secondAttribute, NSLayoutAttribute.centerXWithinMargins)
+        XCTAssertEqual(constraintOne.constant, -10.0)
+        XCTAssertEqual(constraintOne.multiplier, 2.0)
+        XCTAssertEqual(constraintOne.priority, 500)
+
+        XCTAssertEqual(constraintTwo.isActive, true)
+        XCTAssertEqual(constraintTwo.firstItem as! UIView, viewOne)
+        XCTAssertEqual(constraintTwo.firstAttribute, NSLayoutAttribute.centerY)
+        XCTAssertEqual(constraintTwo.relation, NSLayoutRelation.equal)
+        XCTAssertEqual(constraintTwo.secondItem as! UIView, viewTwo)
+        XCTAssertEqual(constraintTwo.secondAttribute, NSLayoutAttribute.centerYWithinMargins)
+        XCTAssertEqual(constraintTwo.constant, -10.0)
         XCTAssertEqual(constraintTwo.multiplier, 2.0)
         XCTAssertEqual(constraintTwo.priority, 500)
     }